:label: [rename] 로그인 response 모델명 수정 (HH-274)
